+++ src/platform/plugins/shared/dashboard/kibana.jsonc
@@ -35,6 +35,7 @@
       "home",
       "spaces",
       "savedObjectsTaggingOss",
+      "savedObjectsTagging",
       "screenshotMode",
       "usageCollection",
       "taskManager",
