+++ x-pack/platform/plugins/shared/fleet/server/services/epm/kibana/assets/saved_object.test.ts
@@ -19,6 +19,9 @@ describe('getSpaceAwareSaveobjectsClients', () => {
     const mockedSavedObjectTagging = {
       createInternalAssignmentService: jest.fn(),
       createTagClient: jest.fn(),
+      getTagsFromReferences: jest.fn(),
+      convertTagNameToId: jest.fn(),
+      replaceTagReferences: jest.fn(),
     };
 
     const scoppedSoClient = savedObjectsClientMock.create();
