+++ x-pack/platform/plugins/shared/saved_objects_tagging/common/references.ts
@@ -6,8 +6,15 @@
  */
 
 import { uniq, intersection } from 'lodash';
-import { SavedObjectReference } from '@kbn/core/types';
+import type {
+  SavedObject,
+  SavedObjectReference,
+  SavedObjectsFindOptionsReference,
+} from '@kbn/core/server';
 import { tagSavedObjectTypeName } from './constants';
+import { Tag } from './types';
+
+type SavedObjectReferenceLike = SavedObjectReference | SavedObjectsFindOptionsReference;
 
 /**
  * Create a {@link SavedObjectReference | reference} for given tag id.
@@ -63,4 +70,44 @@ export const updateTagReferences = ({
   ]);
 
   return [...nonTagReferences, ...newTagIds.map(tagIdToReference)];
+};
+
+export const getTagsFromReferences = (references: SavedObjectReference[], allTags: Tag[]) => {
+  const tagReferences = references.filter((ref) => ref.type === tagSavedObjectTypeName);
+
+  const foundTags: Tag[] = [];
+  const missingRefs: SavedObjectReference[] = [];
+
+  tagReferences.forEach((ref) => {
+    const found = allTags.find((tag) => tag.id === ref.id);
+    if (found) {
+      foundTags.push(found);
+    } else {
+      missingRefs.push(ref);
+    }
+  });
+
+  return {
+    tags: foundTags,
+    missingRefs,
+  };
+};
+
+export const convertTagNameToId = (tagName: string, allTags: Tag[]): string | undefined => {
+  const found = allTags.find((tag) => tag.name.toLowerCase() === tagName.toLowerCase());
+  return found?.id;
+};
+
+export const getObjectTags = (
+  object: { references: SavedObject['references'] },
+  allTags: Tag[]
+) => {
+  return getTagsFromReferences(object.references, allTags);
+};
+
+export const getTag = (tagId: string, allTags: Tag[]): Tag | undefined => {
+  return allTags.find(({ id }) => id === tagId);
+};
+export const getTagIdsFromReferences = (references: SavedObjectReferenceLike[]): string[] => {
+  return references.filter((ref) => ref.type === tagSavedObjectTypeName).map(({ id }) => id);
 };
