+++ x-pack/platform/plugins/shared/saved_objects_tagging/public/ui_api/convert_name_to_reference.ts
@@ -6,8 +6,8 @@
  */
 
 import { SavedObjectsTaggingApiUi } from '@kbn/saved-objects-tagging-oss-plugin/public';
+import { convertTagNameToId } from '../../common';
 import { ITagsCache } from '../services';
-import { convertTagNameToId } from '../utils';
 
 export interface BuildConvertNameToReferenceOptions {
   cache: ITagsCache;
