+++ x-pack/platform/plugins/shared/saved_objects_tagging/public/utils.test.ts
@@ -5,14 +5,7 @@
  * 2.0.
  */
 
-import { SavedObject, SavedObjectReference } from '@kbn/core/types';
-import {
-  getObjectTags,
-  convertTagNameToId,
-  byNameTagSorter,
-  getTagIdsFromReferences,
-  getTag,
-} from './utils';
+import { byNameTagSorter } from './utils';
 
 const createTag = (id: string, name: string = id) => ({
   id,
@@ -20,67 +13,6 @@ const createTag = (id: string, name: string = id) => ({
   description: `desc ${id}`,
   color: '#FFCC00',
   managed: false,
-});
-
-const ref = (type: string, id: string): SavedObjectReference => ({
-  id,
-  type,
-  name: `${type}-ref-${id}`,
-});
-
-const tagRef = (id: string) => ref('tag', id);
-
-const createObject = (refs: SavedObjectReference[]): SavedObject => {
-  return {
-    type: 'unkown',
-    id: 'irrelevant',
-    references: refs,
-  } as SavedObject;
-};
-
-const tag1 = createTag('id-1', 'name-1');
-const tag2 = createTag('id-2', 'name-2');
-const tag3 = createTag('id-3', 'name-3');
-
-const allTags = [tag1, tag2, tag3];
-
-describe('getObjectTags', () => {
-  it('returns the tags for the tag references of the object', () => {
-    const { tags } = getObjectTags(
-      createObject([tagRef('id-1'), ref('dashboard', 'dash-1'), tagRef('id-3')]),
-      allTags
-    );
-
-    expect(tags).toEqual([tag1, tag3]);
-  });
-
-  it('returns the missing references for tags that were not found', () => {
-    const missingRef = tagRef('missing-tag');
-    const refs = [tagRef('id-1'), ref('dashboard', 'dash-1'), missingRef];
-    const { tags, missingRefs } = getObjectTags(createObject(refs), allTags);
-
-    expect(tags).toEqual([tag1]);
-    expect(missingRefs).toEqual([missingRef]);
-  });
-});
-
-describe('convertTagNameToId', () => {
-  it('returns the id for the given tag name', () => {
-    expect(convertTagNameToId('name-2', allTags)).toBe('id-2');
-  });
-
-  it('returns undefined if no tag was found', () => {
-    expect(convertTagNameToId('name-4', allTags)).toBeUndefined();
-  });
-});
-
-describe('getTag', () => {
-  it('returns the tag for the given id', () => {
-    expect(getTag('id-2', allTags)).toEqual(tag2);
-  });
-  it('returns undefined if no tag was found', () => {
-    expect(getTag('id-4', allTags)).toBeUndefined();
-  });
 });
 
 describe('byNameTagSorter', () => {
@@ -95,18 +27,5 @@ describe('byNameTagSorter', () => {
     tags.sort(byNameTagSorter);
 
     expect(tags.map(({ id }) => id)).toEqual(['id-2', 'id-1', 'id-4', 'id-3']);
-  });
-});
-
-describe('getTagIdsFromReferences', () => {
-  it('returns the tag ids from the given references', () => {
-    expect(
-      getTagIdsFromReferences([
-        tagRef('tag-1'),
-        ref('dashboard', 'dash-1'),
-        tagRef('tag-2'),
-        ref('lens', 'lens-1'),
-      ])
-    ).toEqual(['tag-1', 'tag-2']);
   });
 });
