+++ x-pack/platform/plugins/shared/saved_objects_tagging/public/utils.ts
@@ -5,63 +5,12 @@
  * 2.0.
  */
 
-import type { SavedObject, SavedObjectReference } from '@kbn/core/types';
-import type { SavedObjectsFindOptionsReference } from '@kbn/core/public';
 import type { Tag } from '../common/types';
-import { tagSavedObjectTypeName } from '../common';
-
-type SavedObjectReferenceLike = SavedObjectReference | SavedObjectsFindOptionsReference;
-
-export {
-  tagIdToReference,
-  replaceTagReferences as updateTagsReferences,
-} from '../common/references';
-
-export const getObjectTags = (
-  object: { references: SavedObject['references'] },
-  allTags: Tag[]
-) => {
-  return getTagsFromReferences(object.references, allTags);
-};
-
-export const getTagsFromReferences = (references: SavedObjectReference[], allTags: Tag[]) => {
-  const tagReferences = references.filter((ref) => ref.type === tagSavedObjectTypeName);
-
-  const foundTags: Tag[] = [];
-  const missingRefs: SavedObjectReference[] = [];
-
-  tagReferences.forEach((ref) => {
-    const found = allTags.find((tag) => tag.id === ref.id);
-    if (found) {
-      foundTags.push(found);
-    } else {
-      missingRefs.push(ref);
-    }
-  });
-
-  return {
-    tags: foundTags,
-    missingRefs,
-  };
-};
-
-export const convertTagNameToId = (tagName: string, allTags: Tag[]): string | undefined => {
-  const found = allTags.find((tag) => tag.name.toLowerCase() === tagName.toLowerCase());
-  return found?.id;
-};
 
 export const byNameTagSorter = (tagA: Tag, tagB: Tag): number => {
   return tagA.name.localeCompare(tagB.name);
 };
 
-export const getTag = (tagId: string, allTags: Tag[]): Tag | undefined => {
-  return allTags.find(({ id }) => id === tagId);
-};
-
 export const testSubjFriendly = (name: string) => {
   return name.replace(' ', '_');
-};
-
-export const getTagIdsFromReferences = (references: SavedObjectReferenceLike[]): string[] => {
-  return references.filter((ref) => ref.type === tagSavedObjectTypeName).map(({ id }) => id);
 };
