+++ x-pack/platform/plugins/shared/saved_objects_tagging/server/mocks.ts
@@ -13,6 +13,9 @@ const createStartMock = () => {
   const start: jest.Mocked<SavedObjectTaggingStart> = {
     createTagClient: jest.fn(),
     createInternalAssignmentService: jest.fn(),
+    getTagsFromReferences: jest.fn(),
+    convertTagNameToId: jest.fn(),
+    replaceTagReferences: jest.fn(),
   };
 
   start.createTagClient.mockImplementation(() => tagsClientMock.create());
